Mark 1 of Life Mentor
The code will scrape user data from LinkedIn, store it in a Pinecone vector database, and retrieve relevant information when a user asks a career-related question. The system will then analyze the LinkedIn JSON data and generate an appropriate response. PEFT has been executed on the model for fine tuning the model.